user-manager/requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
Django==3.1.4
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
PyJWT==1.7.1
pytz==2023.3
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.4
Unidecode==1.3.6
